Created DockerImageGarbageCollection

+++ docker-api/src/test/java/com/yahoo/vespa/hosted/dockerapi/DockerTest.java
@@ -9,14 +9,13 @@ import org.junit.Test;
 import java.io.IOException;
 import java.net.InetAddress;
 import java.net.URL;
-import java.util.HashSet;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 
-import static junit.framework.TestCase.fail;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 import static org.junit.Assume.assumeTrue;
 
 /**
@@ -33,26 +32,23 @@ import static org.junit.Assume.assumeTrue;
  *     docker stop $(docker ps -a -q)
  *     docker rm $(docker ps -a -q)
  *
- * @author valerijf
+ * @author freva
  * @author dybdahl
  */
 public class DockerTest {
     private DockerImpl docker;
     private static final DockerImage dockerImage = new DockerImage("simple-ipv6-server:Dockerfile");
+    private static final String CONTAINER_NAME_DOCKER_TEST_PREFIX = "docker-test-";
 
     // It is ignored since it is a bit slow and unstable, at least on Mac.
-    @Ignore
     @Test
-    public void testDockerImagePull() throws ExecutionException, InterruptedException {
+    public void testDockerImagePullDelete() throws ExecutionException, InterruptedException {
         DockerImage dockerImage = new DockerImage("busybox:1.24.0");
 
         // Pull the image and wait for the pull to complete
         docker.pullImageAsync(dockerImage).get();
+        assertTrue("Failed to download " + dockerImage.asString() + " image", docker.imageIsDownloaded(dockerImage));
 
-        List<DockerImage> unusedDockerImages = docker.getUnusedDockerImages(new HashSet<>());
-        if (! unusedDockerImages.contains(dockerImage)) {
-            fail("Did not find image as unused, here are all the unused images; " + unusedDockerImages);
-        }
         // Remove the image
         docker.deleteImage(dockerImage);
         assertFalse("Failed to delete " + dockerImage.asString() + " image", docker.imageIsDownloaded(dockerImage));
@@ -60,13 +56,13 @@ public class DockerTest {
 
     // Ignored because the test is very slow (several minutes) when swap is enabled, to disable: (Linux)
     // $ sudo swapoff -a
-    @Ignore
+
     @Test
     public void testOutOfMemoryDoesNotAffectOtherContainers() throws InterruptedException, ExecutionException, IOException {
         String hostName1 = "docker10.test.yahoo.com";
         String hostName2 = "docker11.test.yahoo.com";
-        ContainerName containerName1 = new ContainerName("test-container-1");
-        ContainerName containerName2 = new ContainerName("test-container-2");
+        ContainerName containerName1 = new ContainerName(CONTAINER_NAME_DOCKER_TEST_PREFIX + 1);
+        ContainerName containerName2 = new ContainerName(CONTAINER_NAME_DOCKER_TEST_PREFIX + 2);
         InetAddress inetAddress1 = InetAddress.getByName("172.18.0.10");
         InetAddress inetAddress2 = InetAddress.getByName("172.18.0.11");
 
@@ -98,7 +94,7 @@ public class DockerTest {
 
     @Test
     public void testContainerCycle() throws IOException, InterruptedException, ExecutionException {
-        ContainerName containerName = new ContainerName("foo");
+        ContainerName containerName = new ContainerName(CONTAINER_NAME_DOCKER_TEST_PREFIX + "foo");
         docker.createContainerCommand(dockerImage, containerName, "hostName1").create();
         List<Container> managedContainers = docker.getAllManagedContainers();
         assertThat(managedContainers.size(), is(1));
@@ -125,8 +121,8 @@ public class DockerTest {
     public void testDockerNetworking() throws InterruptedException, ExecutionException, IOException {
         String hostName1 = "docker10.test.yahoo.com";
         String hostName2 = "docker11.test.yahoo.com";
-        ContainerName containerName1 = new ContainerName("test-container-1");
-        ContainerName containerName2 = new ContainerName("test-container-2");
+        ContainerName containerName1 = new ContainerName(CONTAINER_NAME_DOCKER_TEST_PREFIX + 1);
+        ContainerName containerName2 = new ContainerName(CONTAINER_NAME_DOCKER_TEST_PREFIX + 2);
         InetAddress inetAddress1 = InetAddress.getByName("172.18.0.10");
         InetAddress inetAddress2 = InetAddress.getByName("172.18.0.11");
 
@@ -164,8 +160,10 @@ public class DockerTest {
 
         // Clean up any non deleted containers from previous tests
         docker.getAllManagedContainers().forEach(container -> {
-            if (container.isRunning) docker.stopContainer(container.name);
-            docker.deleteContainer(container.name);
+            if (container.name.asString().startsWith(CONTAINER_NAME_DOCKER_TEST_PREFIX)) {
+                if (container.isRunning) docker.stopContainer(container.name);
+                docker.deleteContainer(container.name);
+            }
         });
     }
